src/requirements.txt
aerich==0.6.3
asyncpg==0.25.0
fastapi==0.78.0
gunicorn==20.1.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.9.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
requests==2.28.1
tortoise-orm==0.19.1
uvicorn==0.18.2
